Rename components and directories
Why:
Given that we have removed redux, we can make some file/directory name
changes to make the src folder less nested. This is good practice,
because it allows for shorter relative paths.

This commit:
Renames some files in the src directory.
