+++ src/TimeSelector/TimeSelectorFooter.js
@@ -1,7 +1,7 @@
 import React, { Component } from 'react'
 import { Text, View, StyleSheet } from 'react-native'
 
-import { Colors, Spacing, Typography } from '../../styles'
+import { Colors, Spacing, Typography } from '../styles'
 
 class TimeSelectorFooter extends Component {
   render() {
