+++ src/assets/thoughtbot-assets-pack/README.md
