+++ src/assets/thoughtbot-assets-pack/circle/png/robot_only_default.png
