+++ src/assets/thoughtbot-assets-pack/horizontal/png/horizontal_default.png
