+++ src/assets/thoughtbot-assets-pack/vertical/png/vertical_default.png
